requirements.txt
# Web framework and async support
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Telegram Bot API
python-telegram-bot==21.9

# HTTP client for API calls
httpx==0.28.1

# Configuration and validation
pydantic==2.10.3
pydantic-settings==2.7.0
PyYAML==6.0.2

# Database
aiosqlite==0.20.0

# Logging and monitoring
structlog==25.4.0

# Development and testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Security
python-multipart==0.0.20
